+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Removed the "Check the baseline" task from task lists
 - Conversion projects now have a route of voluntary or sponsored
 - A conversion project can now be added on the same day as the advisory board
+- Projects that have no user assigned to them no longer appear in lists of
+  in-progress projects
 
 ### Added